requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.4
djangorestframework==3.14.0
Markdown==3.5.1
psycopg2-binary